SKT-6:Applied the necessary Middleware implementation by eliminating the redundant portion and modifying it.
