+++ src/app.controller.ts
@@ -1,11 +1,11 @@
-import { Controller, Post } from "@nestjs/common";
+import { Controller, Get } from "@nestjs/common";
 import { AppService } from "./app.service";
 
 @Controller()
 export class AppController {
   constructor(private readonly appService: AppService) {}
 
-  @Post("/")
+  @Get()
   getHello(): string {
     return this.appService.getHello();
   }
