--- src/common/employee/employee.module.ts
@@ -1,7 +1,0 @@
-import { Module } from "@nestjs/common";
-import { RolesController } from "./roles/roles.controller";
-
-@Module({
-  controllers: [RolesController],
-})
-export class EmployeeModule {}
