--- src/common/employee/roles/roles.controller.ts
@@ -1,13 +1,0 @@
-import { Controller, Get } from "@nestjs/common";
-
-@Controller("roles")
-export class RolesController {
-  @Get()
-  roleHistory(): string {
-    return "Roles History Page";
-  }
-  @Get("demo-page")
-  roleHistoryDemo(): string {
-    return "Role History Demo Page";
-  }
-}
