+++ src/common/middleware/logger.middleware.ts
@@ -1,0 +1,12 @@
+/* eslint-disable prettier/prettier */
+import { Injectable, NestMiddleware } from "@nestjs/common";
+import { NextFunction, Request, Response } from "express";
+
+@Injectable()
+export class Logger implements NestMiddleware {
+  use(req: Request, res: Response, next: NextFunction) {
+    console.log("Request logged from global Middleware...", req.url, req.headers);
+
+    next();
+  }
+}
